feat: adds chips in autocomplete multiple (#4345)
* feat: adds chips in autocomplete

* Update packages/eds-core-react/src/components/Autocomplete/utils.ts


---------
